+++ src/dms_provider_bridge/__init__.py
@@ -1,3 +1,3 @@
 __all__ = ["__version__"]
 
-__version__ = "0.8.0-beta"
+__version__ = "0.8.1-beta"